Refactor packages and rename local variable based on feedback from Sonar.
