+++ project-set/components/rate-limiting/src/main/java/com/rackspace/papi/components/ratelimit/RateLimitingRequestInfo.java
@@ -11,8 +11,6 @@ import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 
 public class RateLimitingRequestInfo {
-
-   private static final HeaderValue DEFAULT_EMPTTY = new HeaderValueImpl("", -1.0);
    
    private final HttpMethod requestMethod;
    private final List<? extends HeaderValue> userGroup;
